backend-drf/requirements.txt
# Core Django
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-decouple==3.8
asgiref==3.8.1
sqlparse==0.5.3

# Production server & static files
gunicorn==23.0.0
whitenoise==6.9.0

# Database (PostgreSQL for production)
psycopg2-binary==2.9.9
dj-database-url==2.3.0

# ML inference — onnxruntime reemplaza TensorFlow (~60MB vs ~1GB)
onnxruntime==1.21.0
numpy==2.1.3
scikit-learn==1.6.1
joblib==1.5.1
scipy==1.15.3

# Stock data
pandas==2.2.3
yfinance==0.2.61
yahooquery==2.3.7
alpha-vantage==3.0.0
finnhub-python==2.4.27
beautifulsoup4==4.13.4
requests==2.32.3
multitasking==0.0.11
frozendict==2.4.6
peewee==3.18.1
curl_cffi==0.11.1

# Utilities
Pillow==11.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0
packaging==25.0
six==1.17.0
PyJWT==2.9.0
typing_extensions==4.13.2
setuptools==80.9.0
